LibMedia: Split out Matroska::SampleIterator and Streamer
