+++ Libraries/LibMedia/Containers/Matroska/Reader.cpp
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2021, Hunter Salyer <thefalsehonesty@gmail.com>
- * Copyright (c) 2022-2025, Gregory Bertilson <gregory@ladybird.org>
+ * Copyright (c) 2022-2026, Gregory Bertilson <gregory@ladybird.org>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
@@ -707,81 +707,6 @@ static AK::Duration block_timestamp_to_duration(AK::Duration cluster_timestamp, 
     return cluster_timestamp + AK::Duration::from_nanoseconds(timestamp_offset_in_cluster_offset);
 }
 
-DecoderErrorOr<Vector<ByteBuffer>> SampleIterator::get_frames(Block block)
-{
-    Streamer streamer { m_stream_cursor };
-    TRY(streamer.seek_to_position(block.data_position()));
-    Vector<ByteBuffer> frames;
-
-    if (block.lacing() == Block::Lacing::EBML) {
-        auto frames_start_position = streamer.position();
-        auto frame_count = TRY(streamer.read_octet()) + 1;
-        Vector<u64> frame_sizes;
-        frame_sizes.ensure_capacity(frame_count);
-
-        u64 frame_size_sum = 0;
-        u64 previous_frame_size;
-        auto first_frame_size = TRY(streamer.read_variable_size_integer());
-        frame_sizes.append(first_frame_size);
-        frame_size_sum += first_frame_size;
-        previous_frame_size = first_frame_size;
-
-        for (int i = 0; i < frame_count - 2; i++) {
-            auto frame_size_difference = TRY(streamer.read_variable_size_signed_integer());
-            u64 frame_size;
-            // FIXME: x - (-y) == x + y?
-            if (frame_size_difference < 0)
-                frame_size = previous_frame_size - (-frame_size_difference);
-            else
-                frame_size = previous_frame_size + frame_size_difference;
-            frame_sizes.append(frame_size);
-            frame_size_sum += frame_size;
-            previous_frame_size = frame_size;
-        }
-        frame_sizes.append(block.data_size() - frame_size_sum - (streamer.position() - frames_start_position));
-
-        for (int i = 0; i < frame_count; i++) {
-            // FIXME: ReadonlyBytes instead of copying the frame data?
-            auto current_frame_size = frame_sizes.at(i);
-            frames.append(TRY(streamer.read_raw_octets(current_frame_size)));
-        }
-    } else if (block.lacing() == Block::Lacing::FixedSize) {
-        auto frame_count = TRY(streamer.read_octet()) + 1;
-        auto frames_data_size = block.data_size() - 1;
-        if ((frames_data_size % frame_count) != 0)
-            return DecoderError::corrupted("Block with fixed-size frames has non-divisible size"sv);
-        auto individual_frame_size = frames_data_size / frame_count;
-        for (int i = 0; i < frame_count; i++)
-            frames.append(TRY(streamer.read_raw_octets(individual_frame_size)));
-    } else if (block.lacing() == Block::Lacing::XIPH) {
-        auto frames_start_position = streamer.position();
-
-        auto frame_count_minus_one = TRY(streamer.read_octet());
-        frames.ensure_capacity(frame_count_minus_one + 1);
-
-        auto frame_sizes = Vector<size_t>();
-        frame_sizes.ensure_capacity(frame_count_minus_one);
-        for (auto i = 0; i < frame_count_minus_one; i++) {
-            auto frame_size = 0;
-            while (true) {
-                auto octet = TRY(streamer.read_octet());
-                frame_size += octet;
-                if (octet < 255)
-                    break;
-            }
-            frame_sizes.append(frame_size);
-        }
-
-        for (auto i = 0; i < frame_count_minus_one; i++)
-            frames.append(TRY(streamer.read_raw_octets(frame_sizes[i])));
-        frames.append(TRY(streamer.read_raw_octets(block.data_size() - (streamer.position() - frames_start_position))));
-    } else {
-        frames.append(TRY(streamer.read_raw_octets(block.data_size())));
-    }
-
-    return frames;
-}
-
 static void set_block_duration_to_default(Block& block, TrackBlockContext const& context)
 {
     if (context.default_duration != 0)
@@ -1192,287 +1117,6 @@ DecoderErrorOr<SampleIterator> Reader::seek_to_random_access_point(SampleIterato
 Optional<Vector<TrackCuePoint> const&> Reader::cue_points_for_track(u64 track_number)
 {
     return m_cues.get(track_number);
-}
-
-DecoderErrorOr<Block> SampleIterator::next_block()
-{
-    Streamer streamer { m_stream_cursor };
-    TRY(streamer.seek_to_position(m_position));
-
-    Optional<Block> block;
-
-    while (true) {
-#if MATROSKA_TRACE_DEBUG
-        auto element_position = streamer.position();
-#endif
-        auto element_id = TRY(streamer.read_element_id());
-#if MATROSKA_TRACE_DEBUG
-        dbgln("Iterator found element with ID {:#010x} at offset {} within the segment.", element_id, element_position);
-#endif
-
-        auto maybe_set_block = [&](Block&& candidate_block) {
-            if (candidate_block.track_number() != m_track_number)
-                return;
-            block = move(candidate_block);
-        };
-
-        if (element_id == CLUSTER_ELEMENT_ID) {
-            dbgln_if(MATROSKA_DEBUG, "  Iterator is parsing new cluster.");
-            m_current_cluster = TRY(Reader::parse_cluster_element(streamer, m_segment_timestamp_scale));
-        } else if (element_id == SIMPLE_BLOCK_ID) {
-            if (!m_current_cluster.has_value()) {
-                dbgln("  Iterator encountered a simple block before parsing a Cluster.");
-                TRY(streamer.read_unknown_element());
-            } else {
-                dbgln_if(MATROSKA_TRACE_DEBUG, "  Iterator is parsing a new simple block.");
-                auto candidate_block = TRY(Reader::parse_simple_block(streamer, m_current_cluster->timestamp(), m_segment_timestamp_scale, m_track_block_contexts));
-                maybe_set_block(move(candidate_block));
-            }
-        } else if (element_id == BLOCK_GROUP_ID) {
-            if (!m_current_cluster.has_value()) {
-                dbgln("  Iterator encountered a BlockGroup before parsing a Cluster.");
-                TRY(streamer.read_unknown_element());
-            } else {
-                dbgln_if(MATROSKA_TRACE_DEBUG, "  Iterator is parsing a new block group.");
-                auto candidate_block = TRY(Reader::parse_block_group(streamer, m_current_cluster->timestamp(), m_segment_timestamp_scale, m_track_block_contexts));
-                maybe_set_block(move(candidate_block));
-            }
-        } else if (element_id == SEGMENT_ELEMENT_ID) {
-            dbgln("Malformed file, found a segment element within the root segment element. Jumping into it.");
-            [[maybe_unused]] auto segment_size = TRY(streamer.read_variable_size_integer());
-        } else {
-            dbgln_if(MATROSKA_TRACE_DEBUG, "  Iterator is skipping unknown element with ID {:#010x}.", element_id);
-            TRY(streamer.read_unknown_element());
-        }
-
-        m_position = streamer.position();
-        if (block.has_value()) {
-            m_last_timestamp = block->timestamp();
-            return block.release_value();
-        }
-    }
-
-    VERIFY_NOT_REACHED();
-}
-
-SampleIterator::SampleIterator(NonnullRefPtr<MediaStreamCursor> const& stream_cursor, u64 track_number, TrackBlockContexts&& track_contexts, u64 timestamp_scale, size_t segment_contents_position, size_t position)
-    : m_stream_cursor(stream_cursor)
-    , m_track_number(track_number)
-    , m_track_block_contexts(move(track_contexts))
-    , m_segment_timestamp_scale(timestamp_scale)
-    , m_segment_contents_position(segment_contents_position)
-    , m_position(position)
-{
-}
-
-SampleIterator::~SampleIterator() = default;
-
-DecoderErrorOr<void> SampleIterator::seek_to_cue_point(TrackCuePoint const& cue_point, CuePointTarget target)
-{
-    // This is a private function. The position getter can return optional, but the caller should already know that this track has a position.
-    auto const& cue_position = cue_point.position;
-    Streamer streamer { m_stream_cursor };
-    TRY(streamer.seek_to_position(m_segment_contents_position + cue_position.cluster_position()));
-
-    auto element_id = TRY(streamer.read_element_id());
-    if (element_id != CLUSTER_ELEMENT_ID)
-        return DecoderError::corrupted("Cue point's cluster position didn't point to a cluster"sv);
-
-    m_current_cluster = TRY(Reader::parse_cluster_element(streamer, m_segment_timestamp_scale));
-    dbgln_if(MATROSKA_DEBUG, "SampleIterator set to cue point at timestamp {}ms", m_current_cluster->timestamp().to_milliseconds());
-
-    if (target == CuePointTarget::Cluster) {
-        m_position = streamer.position();
-        m_last_timestamp = m_current_cluster->timestamp();
-    } else {
-        m_position = streamer.position() + cue_position.block_offset();
-        m_last_timestamp = cue_point.timestamp;
-    }
-    return {};
-}
-
-Streamer::Streamer(NonnullRefPtr<MediaStreamCursor> const& stream_cursor)
-    : m_stream_cursor(stream_cursor)
-{
-}
-
-Streamer::~Streamer() = default;
-
-DecoderErrorOr<String> Streamer::read_string()
-{
-    auto string_length = TRY(read_variable_size_integer());
-    auto string_data = TRY(read_raw_octets(string_length));
-    auto const* string_data_raw = reinterpret_cast<char const*>(string_data.data());
-    auto string_value = String::from_utf8(ReadonlyBytes(string_data.data(), strnlen(string_data_raw, string_length)));
-    if (string_value.is_error())
-        return DecoderError::format(DecoderErrorCategory::Invalid, "String is not valid UTF-8");
-    return string_value.release_value();
-}
-
-DecoderErrorOr<u8> Streamer::read_octet()
-{
-    u8 result;
-    Bytes bytes { &result, 1 };
-    TRY(m_stream_cursor->read_into(bytes));
-    return bytes[0];
-}
-
-DecoderErrorOr<i16> Streamer::read_i16()
-{
-    return (TRY(read_octet()) << 8) | TRY(read_octet());
-}
-
-DecoderErrorOr<u32> Streamer::read_element_id()
-{
-    auto length_byte = TRY(read_octet());
-    auto length = count_leading_zeroes_safe(length_byte) + 1;
-    if (length > 4)
-        return DecoderError::corrupted("Element ID must be 4 bytes or less"sv);
-    u32 result = length_byte;
-    auto bytes_left = length;
-    while (--bytes_left > 0) {
-        result <<= 8;
-        result |= TRY(read_octet());
-    }
-
-    return result;
-}
-
-DecoderErrorOr<Optional<size_t>> Streamer::read_element_size()
-{
-    auto length_byte = TRY(read_octet());
-    auto length = count_leading_zeroes_safe(length_byte) + 1;
-    if (length > 8)
-        return DecoderError::corrupted("Element Data Size 8 bytes or less"sv);
-
-    size_t result = length_byte;
-    auto bytes_left = length;
-    while (--bytes_left > 0) {
-        result <<= 8;
-        result |= TRY(read_octet());
-    }
-
-    auto mask = 0xFFFFFFFFFFFFFFFF >> (64 - (7 * length));
-    result &= mask;
-
-    if (result == mask)
-        return OptionalNone();
-
-    return result;
-}
-
-DecoderErrorOr<u64> Streamer::read_variable_size_integer()
-{
-    auto length_byte = TRY(read_octet());
-    auto length = count_leading_zeroes_safe(length_byte) + 1;
-    if (length > 8)
-        return DecoderError::corrupted("VINT is too large"sv);
-
-    auto bytes_left = length;
-    auto result = length_byte & (0xFF >> length);
-    while (--bytes_left > 0) {
-        result <<= 8;
-        result |= TRY(read_octet());
-    }
-    return result;
-}
-
-DecoderErrorOr<i64> Streamer::read_variable_size_signed_integer()
-{
-    auto length_byte = TRY(read_octet());
-    auto length = count_leading_zeroes_safe(length_byte) + 1;
-    if (length > 8)
-        return DecoderError::corrupted("VINT is too large"sv);
-
-    auto bytes_left = length;
-    i64 result = length_byte & (0xFF >> length);
-    while (--bytes_left > 0) {
-        result <<= 8;
-        result |= TRY(read_octet());
-    }
-    result -= AK::exp2<i64>((length * 7) - 1) - 1;
-    return result;
-}
-
-DecoderErrorOr<ByteBuffer> Streamer::read_raw_octets(size_t num_octets)
-{
-    auto result = MUST(ByteBuffer::create_uninitialized(num_octets));
-    auto bytes = result.bytes();
-    TRY(m_stream_cursor->read_into(bytes));
-    return result;
-}
-
-DecoderErrorOr<u64> Streamer::read_u64()
-{
-    auto integer_length = TRY(read_variable_size_integer());
-    if (integer_length == 0)
-        return 0;
-    if (integer_length > 8)
-        return DecoderError::corrupted("Integer Element is too large"sv);
-    u64 result = 0;
-    for (size_t i = 0; i < integer_length; i++) {
-        result <<= 8;
-        result |= TRY(read_octet());
-    }
-    return result;
-}
-
-DecoderErrorOr<i64> Streamer::read_i64()
-{
-    auto integer_length = TRY(read_variable_size_integer());
-    if (integer_length == 0)
-        return 0;
-    if (integer_length > 8)
-        return DecoderError::corrupted("Signed Integer Element is too large"sv);
-    i64 result = 0;
-    for (size_t i = 0; i < integer_length; i++) {
-        result <<= 8;
-        result |= TRY(read_octet());
-    }
-    auto shift = 64 - (static_cast<i64>(integer_length) * 8);
-    result <<= shift;
-    result >>= shift;
-    return result;
-}
-
-DecoderErrorOr<double> Streamer::read_float()
-{
-    auto length = TRY(read_variable_size_integer());
-    if (length == 0)
-        return 0;
-    if (length != 4u && length != 8u)
-        return DecoderError::format(DecoderErrorCategory::Invalid, "Float size must be 4 or 8 bytes");
-
-    union {
-        u64 value;
-        float float_value;
-        double double_value;
-    } read_data;
-    read_data.value = 0;
-    for (size_t i = 0; i < length; i++) {
-        read_data.value = (read_data.value << 8u) + TRY(read_octet());
-    }
-    if (length == 4u)
-        return read_data.float_value;
-    return read_data.double_value;
-}
-
-DecoderErrorOr<void> Streamer::read_unknown_element()
-{
-    auto element_length = TRY(read_variable_size_integer());
-    dbgln_if(MATROSKA_TRACE_DEBUG, "Skipping unknown element of size {}.", element_length);
-    TRY(m_stream_cursor->seek(element_length, AK::SeekMode::FromCurrentPosition));
-    return {};
-}
-
-size_t Streamer::position() const
-{
-    return m_stream_cursor->position();
-}
-
-DecoderErrorOr<void> Streamer::seek_to_position(size_t position)
-{
-    return m_stream_cursor->seek(position, AK::SeekMode::SetPosition);
 }
 
 }
